Adds manager.sh logs command
This simplifies setups where you may want to make an alias to
the `manager.sh` file so you can manage your ark server from any
directory as a server admin. Showing the logs was the only thing
missing from the manager binary, but we can't get the logs from
inside the container so we just short circuit with the
`docker compose logs` command.